pyproject.toml
[tool.poetry]
name = "antlr4-alteryx-functions"
version = "0.1.0"
description = ""
authors = ["Rafael Sales <rafael.sales@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.7"
antlr4-python3-runtime = "^4.13.2"
pytest = "^8.3.3"


[tool.poetry.group.dev.dependencies]
jupyterlab = "^4.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
